Conflicts:
	aula/apps/relacioFamilies/views.py
	aula/utils/views.py
	requirements.txt

+++ aula/apps/extSaga/sincronitzaSaga.py
@@ -63,6 +63,7 @@ def sincronitza(f, user = None):
         a=Alumne()
         a.ralc = ''
         a.telefons = ''
+        dni=''
         #a.tutors = ''
         #a.correu_tutors = ''
 
@@ -123,6 +124,10 @@ def sincronitza(f, user = None):
                 a.rp2_nom = unicode(value,'iso-8859-1')
             if columnName.endswith( u"_ADREÇA" ):
                 a.adreca = unicode(value,'iso-8859-1')
+            if columnName.endswith( u"_CP"):
+                a.cp = unicode(value,'iso-8859-1')
+            if columnName.endswith( u"_DOC. IDENTITAT"):
+                dni = unicode(value,'iso-8859-1')
 
 
         if not (trobatGrupClasse and trobatNom and trobatDataNeixement and trobatRalc):
@@ -130,6 +135,12 @@ def sincronitza(f, user = None):
 
 
         alumneDadesAnteriors = None
+        
+        if not bool(a.ralc) or a.ralc=='':
+            autoRalc, _ = ParametreSaga.objects.get_or_create( nom_parametre = 'autoRalc' )
+            if bool(autoRalc.valor_parametre):
+                a.ralc= autoRalc.valor_parametre + dni[-9:]
+                
         try:
             q_mateix_ralc = Q( ralc = a.ralc ) # & Q(  grup__curs__nivell = a.grup.curs.nivell )
 
@@ -202,6 +213,7 @@ def sincronitza(f, user = None):
                 a.data_alta = date.today()
                 a.motiu_bloqueig = u'No sol·licitat'
                 a.tutors_volen_rebre_correu = False
+                a.foto = alumneDadesAnteriors.foto
             else:
                 a.correu_relacio_familia_pare         = alumneDadesAnteriors.correu_relacio_familia_pare
                 a.correu_relacio_familia_mare         = alumneDadesAnteriors.correu_relacio_familia_mare
@@ -210,6 +222,12 @@ def sincronitza(f, user = None):
                 a.periodicitat_faltes                 = alumneDadesAnteriors.periodicitat_faltes
                 a.periodicitat_incidencies            = alumneDadesAnteriors.periodicitat_incidencies
                 a.tutors_volen_rebre_correu           = alumneDadesAnteriors.tutors_volen_rebre_correu = False
+                a.foto = alumneDadesAnteriors.foto
+
+            manteNom, _ = ParametreSaga.objects.get_or_create( nom_parametre = 'mantenirNom' )
+            
+            if manteNom.valor_parametre=='True':
+                a.nom=alumneDadesAnteriors.nom
 
         a.save()
         nivells.add(a.grup.curs.nivell)
@@ -290,17 +308,18 @@ def sincronitza(f, user = None):
                 impartir__dia_impartir__gte = date.today(),
                 alumne__in = AlumnesDonatsDeBaixa ).delete()
 
-    #Treure'ls de les classes: els canvis de grup   #Todo: només si l'àmbit és grup.
-
-    ambit_no_es_el_grup = Q( impartir__horari__assignatura__tipus_assignatura__ambit_on_prendre_alumnes__in = [ 'C', 'N', 'I' ] )
-    ( ControlAssistencia
-      .objects
-      .filter( ambit_no_es_el_grup )
-      .filter( impartir__dia_impartir__gte = date.today() )
-      .filter( alumne__in = AlumnesCanviatsDeGrup )
-      .delete()
-     )
-
+    manteLlista, _ = ParametreSaga.objects.get_or_create( nom_parametre = 'mantenirLlistes' )
+            
+    if manteLlista.valor_parametre!='True':
+        #Treure'ls de les classes: els canvis de grup   #Todo: només si l'àmbit és grup.
+        ambit_no_es_el_grup = Q( impartir__horari__assignatura__tipus_assignatura__ambit_on_prendre_alumnes__in = [ 'C', 'N', 'I' ] )
+        ( ControlAssistencia
+          .objects
+          .filter( ambit_no_es_el_grup )
+          .filter( impartir__dia_impartir__gte = date.today() )
+          .filter( alumne__in = AlumnesCanviatsDeGrup )
+          .delete()
+         )
 
     #Altes: posar-ho als controls d'assistència de les classes (?????????)
